👷🏻‍♂️ Update Interfaces

+++ src/interfaces/synthetix/IVirtualSynth.sol
@@ -5,13 +5,16 @@ import "./ISynth.sol";
 
 interface IVirtualSynth {
     // Views
-    function balanceOfUnderlying(address account) external view returns (uint);
+    function balanceOfUnderlying(address account)
+        external
+        view
+        returns (uint256);
 
-    function rate() external view returns (uint);
+    function rate() external view returns (uint256);
 
     function readyToSettle() external view returns (bool);
 
-    function secsLeftInWaitingPeriod() external view returns (uint);
+    function secsLeftInWaitingPeriod() external view returns (uint256);
 
     function settled() external view returns (bool);
